flashcard_generation/requirements.txt
xai-sdk>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0

